Conflicts:
	data/mtrl/leaf
	share/glext.h

git-svn-id: https://s.snth.net/svn/neverball/trunk@3607 78b8d119-cf0a-0410-b17c-f493084dd1d7

+++ ball/hud.c
@@ -291,7 +291,7 @@ void hud_speed_pulse(int speed)
 
     for (i = SPEED_NONE + 1; i < SPEED_MAX; i++)
     {
-        const GLfloat *c = gui_gry;
+        const GLubyte *c = gui_gry;
 
         if (speed != SPEED_NONE)
         {
